solutions/changkun: implement quadric error mesh simplification
This change implements a naive edge collapse so that the QEM simplification is complete. Note that this edge collapse has several known issues and might not be resolved in a trivial way. So it might be a good idea to re-implement the entire edge collapse as noted in the KNOWN ISSUES comments.

Nevertheless, the current version works with closed meshes.